refactor: AnimatePresence를 SubCategoryFilter에서 RenderTabContent로 이동하여 서브카테고리 변경 시 애니메이션 정상 동작
- 기존 SubCategoryFilter 내부에서 AnimatePresence를 사용하던 로직을 제거
- RenderTabContent에서 key를 바꿔주며 컴포넌트가 올바르게 언마운트/마운트되도록 수정
